+++ CHANGELOG.md
@@ -5,6 +5,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## [Unreleased] -->
 
+## [0.8.1] - 2022-05-26
+
+### Fixed
+
+- removed low resolution that generated offset sometimes
+- fix download on Linux
+
 ## [0.8.0] - 2022-05-25
 
 ### Added